requirements.txt
aiohttp
async-timeout
certifi
python-dotenv
mysql-connector-python
pytz
requests
websocket-client
python-dateutil
multidict
yarl
charset-normalizer
idna
typing-extensions
attrs
frozenlist
aiohttp-retry
asyncio